# apps/docs/openapi/writing-app-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Writing App API",
    "description": "개발 문서",
    "version": "0.0.1"
  },
  "paths": {
...
    "/courses/{courseId}/progress": {
      "get": {
        "operationId": "getCoursesByCourseIdProgress",
        "responses": {
          "200": {
            "description": "현재 학습자의 코스 진행입니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "completedCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "courseId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "nextLessonId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "progressPercent": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 100
                    },
                    "totalLessons": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "completedCount",
                    "courseId",
                    "progressPercent",
                    "totalLessons"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "unauthorized"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "404": {
            "description": "코스를 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "course-not-found"
                    },
                    "message": {
                      "type": "string",
                      "const": "코스를 찾을 수 없습니다."
                    },
                    "courseId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message", "courseId"]
                }
              }
            }
          },
          "500": {
            "description": "콘텐츠 시드가 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-content-seed"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "503": {
            "description": "데이터베이스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "database-unavailable"
                    },
                    "message": {
                      "type": "string",
                      "const": "데이터베이스를 사용할 수 없습니다."
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "courseId",
            "required": true
          }
        ]
      }
    },
    "/lessons/{lessonId}/progress": {
      "get": {
        "operationId": "getLessonsByLessonIdProgress",
        "responses": {
          "200": {
            "description": "현재 학습자의 레슨 진행입니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "answers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "answer": {
                            "type": "string"
                          },
                          "stepId": {
                            "type": "string",
                            "minLength": 1
                          }
                        },
                        "required": ["answer", "stepId"]
                      }
                    },
                    "currentStepId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "string",
                      "enum": ["not-started", "in-progress", "completed"]
                    },
                    "stepOrder": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "answers",
                    "currentStepId",
                    "lessonId",
                    "status",
                    "stepOrder"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "unauthorized"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "404": {
            "description": "레슨을 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "lesson-not-found"
                    },
                    "message": {
                      "type": "string",
                      "const": "레슨을 찾을 수 없습니다."
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message", "lessonId"]
                }
              }
            }
          },
          "500": {
            "description": "콘텐츠 시드가 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-content-seed"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "503": {
            "description": "데이터베이스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "database-unavailable"
                    },
                    "message": {
                      "type": "string",
                      "const": "데이터베이스를 사용할 수 없습니다."
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "lessonId",
            "required": true
          }
        ]
      },
      "put": {
        "operationId": "putLessonsByLessonIdProgress",
        "responses": {
          "200": {
            "description": "레슨 진행을 저장했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "answers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "answer": {
                            "type": "string"
                          },
                          "stepId": {
                            "type": "string",
                            "minLength": 1
                          }
                        },
                        "required": ["answer", "stepId"]
                      }
                    },
                    "currentStepId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "string",
                      "enum": ["not-started", "in-progress", "completed"]
                    },
                    "stepOrder": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "answers",
                    "currentStepId",
                    "lessonId",
                    "status",
                    "stepOrder"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "요청이 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-request"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "unauthorized"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "404": {
            "description": "레슨을 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "lesson-not-found"
                    },
                    "message": {
                      "type": "string",
                      "const": "레슨을 찾을 수 없습니다."
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message", "lessonId"]
                }
              }
            }
          },
          "500": {
            "description": "콘텐츠 시드가 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-content-seed"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "503": {
            "description": "데이터베이스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "database-unavailable"
                    },
                    "message": {
                      "type": "string",
                      "const": "데이터베이스를 사용할 수 없습니다."
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "lessonId",
            "required": true
          }
        ]
      }
    },
    "/lessons/{lessonId}/answers": {
      "put": {
        "operationId": "putLessonsByLessonIdAnswers",
        "responses": {
          "200": {
            "description": "레슨 답변을 저장했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "saved": {
                      "type": "boolean",
                      "const": true
                    }
                  },
                  "required": ["saved"]
                }
              }
            }
          },
          "400": {
            "description": "요청이 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-request"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "unauthorized"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "404": {
            "description": "레슨을 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "lesson-not-found"
                    },
                    "message": {
                      "type": "string",
                      "const": "레슨을 찾을 수 없습니다."
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message", "lessonId"]
                }
              }
            }
          },
          "500": {
            "description": "콘텐츠 시드가 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-content-seed"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "503": {
            "description": "데이터베이스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "database-unavailable"
                    },
                    "message": {
                      "type": "string",
                      "const": "데이터베이스를 사용할 수 없습니다."
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "lessonId",
            "required": true
          }
        ]
      }
    },
    "/lessons/{lessonId}/complete": {
      "post": {
        "operationId": "postLessonsByLessonIdComplete",
        "responses": {
          "200": {
            "description": "레슨을 완료했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "completedAt": {
                      "type": "string"
                    },
                    "completedCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "string",
                      "const": "completed"
                    },
                    "wasAlreadyCompleted": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "completedAt",
                    "completedCount",
                    "lessonId",
                    "status",
                    "wasAlreadyCompleted"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "unauthorized"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "404": {
            "description": "레슨을 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "lesson-not-found"
                    },
                    "message": {
                      "type": "string",
                      "const": "레슨을 찾을 수 없습니다."
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message", "lessonId"]
                }
              }
            }
          },
          "500": {
            "description": "콘텐츠 시드가 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-content-seed"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "503": {
            "description": "데이터베이스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "database-unavailable"
                    },
                    "message": {
                      "type": "string",
                      "const": "데이터베이스를 사용할 수 없습니다."
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "lessonId",
            "required": true
          }
        ]
      }
    },
...
  },
...
}